projects/emailguard/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
dnspython==2.5.0
email-validator==2.1.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0

# Development dependencies
pytest==7.4.4
pytest-asyncio==0.23.3